Add derived type query test cases
